requirements.txt
aniso8601
beautifulsoup4
boto3
botocore
bs4
cachetools
cairocffi
CairoSVG
certifi
chardet
Click
colorama
cryptography
cssselect2
defusedxml
docutils
Events
Flask
Flask-Cors
Flask-RESTful
Flask-SQLAlchemy
funcy
html5lib
idna
isort
itsdangerous
Jinja2
jmespath
jwt
lazy-object-proxy
MarkupSafe
mccabe
mysqlclient
Naked
pandas
passlib
pdfkit
Pillow
pycparser
pycryptodome
pylint
pymemcache
Pyphen
python-barcode
python-dateutil
python-memcached
pytz
PyYAML
redis
reportlab
requests
s3transfer
shellescape
six
soupsieve
SQLAlchemy
tinycss2
urllib3
psycopg2
python-oauth2
firebase-admin
flask 
flask-restful 
pyrebase4

